move MultiSearchError into separate file, rename error
